🙅‍♂️ Add and use eleventyignore file
When `watch`ing, Eleventy would build on the save of the Sass file,
which meant that the changes to the CSS hadn't been compiled and saved
ready to be used when the site refreshed.

Eleventy uses the `.gitignore` file to see what it should ignore. But we
need to commit the Sass files to git.

Enter `.eleventyignore`, which can be used as well as gitignore, to tell
Eleventy which files to ignore. But then Eleventy still ignores the
CSS files, since they're being ignored by git.

So we have to tell Eleventy to exclusively use the `.eleventyignore`
file - this means some duplication, but as these ignore files are rarely
updated it shouldn't be a problem - but it does mean we can clearly tell
git what to ignore and Eleventy what to ignore separately.