🎨 Update theme colour
Theme colour used to be yellow - but that colour isn't used any more, so is
replaced with the same colour as the border colour. Added the correct colour for
both dark and light mode.

Required a bit of tweaking to work on Safari in iOS, due to the notch affecting
the spacing on the sides of the page when in landscape mode.